+++ Ash/src/Vulkan/Window/Window.cpp
@@ -35,6 +35,6 @@ namespace Ash::Vulkan
 		Width = width;
 		Height = height;
 
-		LastResizeTime = glfwGetTime();
+		LastResizeTime = (float)glfwGetTime();
 	}
 }
